+++ assets/src/components/account/Account.tsx
@@ -5,21 +5,26 @@ import {
   TabPanel,
   useSetBreadcrumbs,
 } from '@pluralsh/design-system'
-import { useContext, useRef } from 'react'
+import { useRef } from 'react'
 import { Link, Outlet, useLocation } from 'react-router-dom'
 
 import { ResponsiveLayoutSidecarContainer } from 'components/utils/layout/ResponsiveLayoutSidecarContainer'
 import { ResponsiveLayoutSpacer } from 'components/utils/layout/ResponsiveLayoutSpacer'
 import { ResponsiveLayoutContentContainer } from 'components/utils/layout/ResponsiveLayoutContentContainer'
 import { ResponsiveLayoutSidenavContainer } from 'components/utils/layout/ResponsiveLayoutSidenavContainer'
-import { LoginContext } from 'components/contexts'
+import { useLogin } from 'components/contexts'
 import { ResponsiveLayoutPage } from 'components/utils/layout/ResponsiveLayoutPage'
 import { useTheme } from 'styled-components'
-import { ConsoleConfiguration, User } from 'generated/graphql'
+import {
+  ConsoleConfiguration,
+  PersonaConfigurationFragment,
+  User,
+} from 'generated/graphql'
 
 const getDirectory = (
   me: User,
-  configuration: Nullable<ConsoleConfiguration>
+  configuration: Nullable<ConsoleConfiguration>,
+  personaConfiguration: Nullable<PersonaConfigurationFragment>
 ) => [
   { path: 'users', label: 'Users', enabled: true },
   { path: 'groups', label: 'Groups', enabled: true },
@@ -37,6 +42,13 @@ const getDirectory = (
     enabled: me?.roles?.admin && configuration?.gitStatus?.cloned,
   },
   { path: 'settings', label: 'Account settings', enabled: true },
+  {
+    path: 'audits',
+    label: 'Audits',
+    enabled: !!(
+      personaConfiguration?.all || personaConfiguration?.sidebar?.audits
+    ),
+  },
 ]
 
 const breadcrumbs: Breadcrumb[] = [{ label: 'account', url: '/account' }]
@@ -44,7 +56,7 @@ const breadcrumbs: Breadcrumb[] = [{ label: 'account', url: '/account' }]
 export default function Account() {
   const theme = useTheme()
   const tabStateRef = useRef<any>(null)
-  const { me, configuration } = useContext(LoginContext)
+  const { me, configuration, personaConfiguration } = useLogin()
   const { pathname } = useLocation()
   const pathPrefix = '/account'
 
@@ -52,9 +64,11 @@ export default function Account() {
 
   if (!me) return null
 
-  const directory = getDirectory(me, configuration).filter(
-    ({ enabled }) => enabled
-  )
+  const directory = getDirectory(
+    me,
+    configuration,
+    personaConfiguration
+  ).filter(({ enabled }) => enabled)
   const currentTab = directory.find(
     (tab) => pathname?.startsWith(`${pathPrefix}/${tab.path}`)
   )
